[FEAT] packaging, and create settingactivity + layout
