+++ .idea/misc.xml
@@ -58,6 +58,7 @@
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/font/noto_sans_kr.xml" value="0.14192708333333334" />
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/activity_home.xml" value="0.264" />
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/activity_setting.xml" value="0.4" />
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/activity_sign_in.xml" value="0.264" />
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/activity_sign_up.xml" value="0.11510416666666666" />
         <entry key="..\:/Users/gua05/AndroidStudioProjects/3rd_seminar/app/src/main/res/layout/fragment_camera.xml" value="0.1" />
