+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,13 @@
         android:theme="@style/Theme.3rd_seminar"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".SignInActivity"
+            android:name=".ui.setting.SettingActivity"
             android:exported="false" />
         <activity
-            android:name=".SignUpActivity"
+            android:name=".ui.signin.SignInActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".ui.home.HomeActivity"
