+++ app/src/main/java/com/example/a3rd_seminar/ui/home/sub_fragments/ProfileFragment.kt
@@ -5,8 +5,8 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.a3rd_seminar.ui.home.sub_fragments.profile_fragment.ProfileFollowerFragment
-import com.example.a3rd_seminar.ui.home.sub_fragments.profile_fragment.ProfileRepositoryFragment
+import com.example.a3rd_seminar.ui.home.sub_fragments.profile.ProfileFollowerFragment
+import com.example.a3rd_seminar.ui.home.sub_fragments.profile.ProfileRepositoryFragment
 import com.example.a3rd_seminar.R
 import com.example.a3rd_seminar.databinding.FragmentProfileBinding
