+++ app/src/main/java/com/example/a3rd_seminar/ui/home/sub_fragments/profile/ProfileFollowerFragment.kt
@@ -1,4 +1,4 @@
-package com.example.a3rd_seminar.ui.home.sub_fragments.profile_fragment
+package com.example.a3rd_seminar.ui.home.sub_fragments.profile
 
 import android.os.Bundle
 import android.util.Log
