+++ app/src/main/res/layout/activity_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FEFEFE"
-    tools:context=".SignInActivity">
+    tools:context=".ui.signin.SignInActivity">
 
     <View
         android:id="@+id/img_github"
