+++ app/src/main/res/navigation/nav_sample.xml
@@ -32,7 +32,7 @@
     </fragment>
     <activity
         android:id="@+id/signInActivity"
-        android:name="com.example.a3rd_seminar.SignInActivity"
+        android:name="com.example.a3rd_seminar.ui.signin.SignInActivity"
         android:label="activity_sign_in"
         tools:layout="@layout/activity_sign_in" />
